requirements.txt
yfinance==0.2.33
requests==2.31.0
pandas==2.1.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
openai==1.6.1
pytest==7.4.3
